CheckWorkerHealth in Parallel.Wait() to prevent deadlock
If, for some reason, one or more workers crash and either parallelfox.exe or a thread stops working, the Parallel.Wait() call in the main process will continue waiting for that "busy" worker without realizing that it has crashed. This results in a deadlock. To prevent this, we now check both the Out-of-Process and In-Process workers to ensure they are still alive.